Move chemistry examples
