+++ examples/chemistry/tests/test_docs.py
