+++ examples/chemistry/tests/test_notebooks.py
