+++ examples/chemistry/tests/test_remineralization.py
@@ -111,7 +111,7 @@ from idaes.core.util.model_statistics import degrees_of_freedom
 import idaes.logger as idaeslog
 
 # Import scaling helper functions
-from watertap.examples.chemistry.chem_scaling_utils import (
+from chem_scaling_utils import (
     _set_inherent_rxn_scaling,
     _set_rate_rxn_scaling,
     _set_mat_bal_scaling_FTPx,
