+++ examples/chemistry/tests/test_tutorials.py
