+++ examples/chemistry/tests/test_water_softening.py
